refactor: useSavingsCalculator 훅의 입력값과 상품목록을 관심사 분리 (#1)
